Removed unused imports Reformatted and rearranged the code

+++ src/main/java/com/codecool/mhmm/stickman/game_objects/characters/enemy/Skeleton.java
@@ -1,12 +1,13 @@
 package com.codecool.mhmm.stickman.game_objects.characters.enemy;
 
 import com.codecool.mhmm.stickman.game_objects.GameObjectType;
+
 import javax.persistence.Entity;
 
 @Entity
 public class Skeleton extends Enemy {
-    public Skeleton(int X, int Y, int hitPoints, int level){
-        super(X, Y, hitPoints, Math.round(level * (3/2)), level);
+    public Skeleton(int X, int Y, int hitPoints, int level) {
+        super(X, Y, hitPoints, (int) Math.round(level * (3.0 / 2.0)), level);
         this.dodgeChance = 25;
         this.type = GameObjectType.SKELETON;
     }
